+++ tools/Azure.Mcp.Tools.Extension/src/ExtensionSetup.cs
@@ -3,7 +3,10 @@
 
 using Azure.Mcp.Core.Areas;
 using Azure.Mcp.Core.Commands;
+using Azure.Mcp.Core.Extensions;
+using Azure.Mcp.Core.Services.Http;
 using Azure.Mcp.Tools.Extension.Commands;
+using Azure.Mcp.Tools.Extension.Services;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Azure.Mcp.Tools.Extension;
@@ -14,7 +17,10 @@ public sealed class ExtensionSetup : IAreaSetup
 
     public void ConfigureServices(IServiceCollection services)
     {
+        services.AddHttpClientServices();
+        services.AddSingleton<ICliGenerateService, CliGenerateService>();
         services.AddSingleton<AzqrCommand>();
+        services.AddSingleton<CliGenerateCommand>();
     }
 
     public CommandGroup RegisterCommands(IServiceProvider serviceProvider)
@@ -26,6 +32,10 @@ public sealed class ExtensionSetup : IAreaSetup
         var azqr = serviceProvider.GetRequiredService<AzqrCommand>();
         extension.AddCommand(azqr.Name, azqr);
 
+        var cli = new CommandGroup("cli", "Commands for helping users to use CLI tools for Azure services operations. Includes operations for generating Azure CLI commands.");
+        extension.AddSubGroup(cli);
+        var cliGenerateCommand = serviceProvider.GetRequiredService<CliGenerateCommand>();
+        cli.AddCommand(cliGenerateCommand.Name, cliGenerateCommand);
         return extension;
     }
 }